% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FireStore.R
\name{fireStore.download}
\alias{fireStore.download}
\title{Download allows us to download a whole collection of Cloud Firestore}
\usage{
fireStore.download(project_id, fileName, token, save_path = NULL)
}
\arguments{
\item{project_id}{This is the unique identifier of your Firebase project {string}}

\item{fileName}{Name of the collection {string}}

\item{token}{token that returns us the funcion FireStore.google_firestore {string}}

\item{save_path}{path to save to hard drive in RDS format {string}}
}
\value{
a dataframe returns with all the information of the collection
}
\description{
fireStore::fireStore.download a dataframe returns with all the data of the collection
}
\examples{
\dontrun{
fireStore.download("mi-proyecto-6465" , "micoleccion", "KxwWNTVdplXFRZwGMkH")
}
}
\author{
Luis Rodriguez
}
