% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FireStore.R
\name{fireStore.count}
\alias{fireStore.count}
\title{Count nos permite hace un conteo de registros de una coleccion  de Cloud Firestore}
\usage{
fireStore.count(project_id, fileName, token)
}
\arguments{
\item{project_id}{Este es el identificador único de tu proyecto de Firebase. {string}}

\item{fileName}{nombre de la coleccion. {string}}

\item{token}{token que nos retorna la funcion FireStore.google_firestore {string}}
}
\value{
retorna un numero de la cantidad de registros
}
\description{
fireStore::fireStore.count nos retorna un numero de la cantidad de registros
}
\examples{
\dontrun{
fireStore.count("mi-proyecto-6465" , "micoleccion", "KxwWNTVdplXFRZwGMkH")
}
}
\author{
Luis Rodriguez
}
