% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FireStore.R
\name{fireStore.download}
\alias{fireStore.download}
\title{Download nos permite descargar toda una coleccion  de Cloud Firestore}
\usage{
fireStore.download(project_id, fileName, token, save_path = NULL)
}
\arguments{
\item{project_id}{Este es el identificador unico de tu proyecto de Firebase. {string}}

\item{fileName}{nombre de la coleccion. {string}}

\item{token}{token que nos retorna la funcion FireStore.google_firestore {string}}

\item{save_path}{ruta para guardar en el disco duro en formato RDS {string}}
}
\value{
retorna un dataframe con el toda la informacion de la coleccion
}
\description{
fireStore::fireStore.download nos retorna un dataframe con la toda la data de la coleccion
}
\examples{
\dontrun{
fireStore.download("mi-proyecto-6465" , "micoleccion", "KxwWNTVdplXFRZwGMkH")
}
}
\author{
Luis Rodriguez
}
