% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FireStore.R
\name{fireStore.google_firestore}
\alias{fireStore.google_firestore}
\title{The OAuth function to get access to the cloud firestore:}
\usage{
fireStore.google_firestore(
  web_client_id = "prompt",
  web_client_secret = "prompt",
  cache = FALSE
)
}
\arguments{
\item{web_client_id}{The Web Client ID of your Google OAuth in your Firebase. {string}}

\item{web_client_secret}{The Web Client Secret of your Google OAuth in your Firebase. {string}}

\item{cache}{Cache the tokens in the .httr-oauth file or not. {boolean}}
}
\value{
Returns the token data.
}
\description{
fireStore::fireStore.google_firestore retrieves a token with read access to the cloud firestore
}
\examples{
\dontrun{
fireStore.google_firestore("xxxxxxxxxxxxx","xxxxxxxxxxxxxxxxx")
}
}
\author{
Paul Spende
}
